qemu_domainjob: Make copy of owner API
Using the job owner API name directly works fine as long as it is a
static string or the owner's thread is still running. However, this is
not always the case. For example, when the owner API name is filled in a
job when we're reconnecting to existing domains after daemon restart,
the dynamically allocated owner name will disappear with the
reconnecting thread. Any follow up usage of the pointer will read random
memory.

Signed-off-by: Jiri Denemark <jdenemar@redhat.com>
Message-Id: <718bfef15fdf80b56a0b8f6e3f1370497a6f9309.1614191420.git.jdenemar@redhat.com>
Reviewed-by: Peter Krempa <pkrempa@redhat.com>