lxc: Including CRIU functions and functions to support C/R.
This patch adds the source code of helper functions into files
lxc_criu.{h,c} to support LXC checkpoint/restore using CRIU binary. To save
container state, LXC follows the same pattern of QEMU and libxl using a file
with a header with metadata, but as CRIU saves multiple files, it needs to
inserted in a unique file using a type of compression. Using TAR for
instance.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>
Message-Id: <20210227040635.73934-3-jcfaracco@gmail.com>