virtmp: Fix @path handling in virTPMEmulatorInit()
This function finds "swtmp", "swtpm_setup" and "swtpm_ioctl"
binaries in $PATH and stores resolved paths in global variables
so that they can be obtainer later. Anyway, the resolved path is
marked as g_autofree and to avoid its freeing later on in the
function the variable is set to NULL manually. Well, we have
g_steal_pointer() for thath.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <e69f9d248b21f0c89562fcd017d14542c240acde.1614617800.git.mprivozn@redhat.com>