feat(usb-hostdev): add ability to filter by USB serial strings too
This patch adds the ability to select a specific USB device when
multiple are available with the same vendor and product IDs.

As an example, I'm using this to pass through a specific USB mouse to my
guest VM. I happen to have two of the exact same model (Logitech MX518),
differing only in their serial number, and only wish to pass through one
of them to my VM.

An example of the config snippet I'm using with this:

    <hostdev mode='subsystem' type='usb' managed='yes'>
      <source>
        <vendor id='0x046d'/>
        <product id='0xc08e'/>
        <serial>0D3087538362</serial>
      </source>
    </hostdev>

Without this patch I would either get an error message on attempted
startup, complaining about having found multiple devices with the same
product or vendor id. Or alternatively I'd have to set the bus number
and device number. Unfortunately the latter tends to change with every
reboot, and in my case whenever my monitor comes out of standby (I'm
using its builtin USB hub).

Message-Id: <20210302162326.3998683-1-giel@mortis.eu>