vircgroup: enforce range limit for cpu.shares
Before the conversion to using systemd DBus API to set the cpu.shares
there was some magic conversion done by kernel which was documented in
virsh manpage as well. Now systemd errors out if the value is out of
range.

Since we enforce the range for other cpu cgroup attributes 'quota' and
'period' it makes sense to do the same for 'shares' as well.

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>
Message-Id: <3d9ea1f927711e7627705dea5f20a6009e6c4320.1614778388.git.phrdina@redhat.com>
Reviewed-by: Michal Privoznik <mprivozn@redhat.com>