qemu_shim: Replace g_file_get_contents() with virFileReadAll()
The qemu_shim (compiled into virt-qemu-run-binary) reads several
files provided by user (XML definition of secret, value of the
secret, XML definition of domain) and it does so using
g_file_get_contents(). This is potentially dangerous, because
there is no limit on the size of files/buffers.

Since this is a standalone binary it's not critical as it can't
cause libvirtd to be OOM killed, but it's still worth fixing as I
am planning on discouraging people from using the GLib function.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <308f65cc9f839b32d6e225beec9e05ae1102ed4c.1615299876.git.mprivozn@redhat.com>